PyTest DQ Framework/requirements.txt
pytest
pytest-html
pandas
sqlalchemy
psycopg2-binary
pyarrow
